Moved library configuration from lib/ to config/
