+++ src/api/eservice/eservice.services.ts
@@ -1,4 +1,4 @@
-import axiosInstance from '@/lib/axios'
+import axiosInstance from '@/config/axios'
 import { BACKEND_FOR_FRONTEND_URL, CATALOG_PROCESS_URL } from '@/config/env'
 import {
   EServiceGetListFlatResponse,
