+++ src/api/react-query-wrappers/__tests__/useQueryWrapper.test.tsx
@@ -1,15 +1,15 @@
 import React from 'react'
 import { renderHook, screen } from '@testing-library/react'
 import { useQueryWrapper } from '../useQueryWrapper'
-import { QueryClientProvider } from '@tanstack/react-query'
-import { queryClient } from '@/lib/react-query'
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 import { vi } from 'vitest'
 import * as hooks from '@/hooks/useJwt'
 import { act } from 'react-dom/test-utils'
 import { ErrorBoundary } from 'react-error-boundary'
 import { rest } from 'msw'
 import { setupServer } from 'msw/node'
-import axiosInstance from '@/lib/axios'
+import axiosInstance from '@/config/axios'
+import { queryClientConfig } from '@/config/query-client'
 
 const server = setupServer(
   rest.get('/test-success', (_, res, ctx) => {
@@ -44,6 +44,8 @@ async function promiseResolvedMock() {
 async function promiseRejectedMock() {
   return await axiosInstance.get('/test-404')
 }
+
+const queryClient = new QueryClient(queryClientConfig)
 
 const wrapper = ({ children }: { children: React.ReactNode }) => (
   <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
