+++ src/contexts/ProvidersWrapper.tsx
@@ -3,9 +3,11 @@ import { ThemeProvider } from '@mui/material'
 import { theme } from '@/config/theme'
 import { ToastNotificationContextProvider } from './ToastNotificationContext'
 import { LoadingOverlayContextProvider } from './LoadingOverlayContext'
-import { QueryClientProvider } from '@tanstack/react-query'
-import { queryClient } from '@/lib/react-query'
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools'
+import { queryClientConfig } from '@/config/query-client'
+
+const queryClient = new QueryClient(queryClientConfig)
 
 const ProvidersWrapper: React.FC<{ children: React.ReactNode }> = ({ children }) => {
   return (
