+++ src/main.tsx
@@ -2,6 +2,6 @@ import React from 'react'
 import ReactDOM from 'react-dom/client'
 import App from './App'
 import '@/index.css'
-import '@/lib/react-i18next'
+import '@/config/react-i18next'
 
 ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(<App />)
